+++ argopy/tests/test_fetchers_data_localftp.py
@@ -1,15 +1,3 @@
-#!/bin/env python
-# -*coding: UTF-8 -*-
-#
-# Test data fetchers
-#
-# This is not designed as it should
-# We need to have:
-# - one class to test the facade API
-# - one class to test specific methods of each backends
-#
-# At this point, we are testing real data fetching both through facade and through direct call to backends.
-
 import os
 import numpy as np
 import xarray as xr
@@ -17,18 +5,17 @@ import shutil
 
 import pytest
 import unittest
-from unittest import TestCase
 
 import argopy
 from argopy import DataFetcher as ArgoDataFetcher
-from argopy.errors import InvalidFetcher, ErddapServerError, CacheFileNotFound, FileSystemHasNoCache, FtpPathError
-from argopy.utilities import list_available_data_src, isconnected, isAPIconnected, erddap_ds_exists
+from argopy.errors import CacheFileNotFound, FileSystemHasNoCache, FtpPathError
+from argopy.utilities import list_available_data_src, is_list_of_strings
 
 AVAILABLE_SOURCES = list_available_data_src()
 
 
 @unittest.skipUnless('localftp' in AVAILABLE_SOURCES, "requires localftp data fetcher")
-class Backend(TestCase):
+class Backend(unittest.TestCase):
     """ Test main API facade for all available dataset and access points of the localftp fetching backend """
     src = 'localftp'
     testcachedir = os.path.expanduser(os.path.join("~", ".argopytest_tmp"))
@@ -89,44 +76,38 @@ class Backend(TestCase):
 
     def __testthis_profile(self, dataset):
         with argopy.set_options(local_ftp=self.local_ftp):
+            fetcher_args = {"src": self.src, 'ds': dataset}
             for arg in self.args['profile']:
                 try:
-                    ds = ArgoDataFetcher(src=self.src, ds=dataset).profile(*arg).to_xarray()
-                    assert isinstance(ds, xr.Dataset)
-                except ErddapServerError:
-                    # Test is passed when something goes wrong because of the erddap server, not our fault !
-                    pass
+                    f = ArgoDataFetcher(**fetcher_args).profile(*arg)
+                    assert is_list_of_strings(f.fetcher.uri)
+                    assert isinstance(f.to_xarray(), xr.Dataset)
                 except Exception:
-                    print("ERROR LOCALFTP request:\n",
-                          ArgoDataFetcher(src=self.src, ds=dataset).profile(*arg).fetcher.uri)
+                    print("ERROR LOCALFTP request:\n", f.fetcher.uri)
                     pass
 
     def __testthis_float(self, dataset):
         with argopy.set_options(local_ftp=self.local_ftp):
+            fetcher_args = {"src": self.src, 'ds': dataset}
             for arg in self.args['float']:
                 try:
-                    ds = ArgoDataFetcher(src=self.src, ds=dataset).float(arg).to_xarray()
-                    assert isinstance(ds, xr.Dataset)
-                except ErddapServerError:
-                    # Test is passed when something goes wrong because of the erddap server, not our fault !
-                    pass
+                    f = ArgoDataFetcher(**fetcher_args).float(arg)
+                    assert is_list_of_strings(f.fetcher.uri)
+                    assert isinstance(f.to_xarray(), xr.Dataset)
                 except Exception:
-                    print("ERROR LOCALFTP request:\n",
-                          ArgoDataFetcher(src=self.src, ds=dataset).float(arg).fetcher.uri)
+                    print("ERROR LOCALFTP request:\n", f.fetcher.uri)
                     pass
 
     def __testthis_region(self, dataset):
         with argopy.set_options(local_ftp=self.local_ftp):
+            fetcher_args = {"src": self.src, 'ds': dataset}
             for arg in self.args['region']:
                 try:
-                    ds = ArgoDataFetcher(src=self.src, ds=dataset).region(arg).to_xarray()
-                    assert isinstance(ds, xr.Dataset)
-                except ErddapServerError:
-                    # Test is passed when something goes wrong because of the erddap server, not our fault !
-                    pass
+                    f = ArgoDataFetcher(**fetcher_args).region(arg)
+                    assert is_list_of_strings(f.fetcher.uri)
+                    assert isinstance(f.to_xarray(), xr.Dataset)
                 except Exception:
-                    print("ERROR LOCALFTP request:\n",
-                          ArgoDataFetcher(src=self.src, ds=dataset).region(arg).fetcher.uri)
+                    print("ERROR LOCALFTP request:\n", f.fetcher.uri)
                     pass
 
     def __testthis(self, dataset):
@@ -156,6 +137,51 @@ class Backend(TestCase):
         self.args['region'] = [[-60, -40, 40., 60., 0., 100.],
                                [-60, -40, 40., 60., 0., 100., '2007-08-01', '2007-09-01']]
         self.__testthis('phy')
+
+
+class BackendParallel(unittest.TestCase):
+    """ This test backend for parallel requests """
+
+    src = "localftp"
+    local_ftp = argopy.tutorial.open_dataset('localftp')[0]
+
+    requests = {}
+    requests["region"] = [
+        [-60, -40, 40., 60., 0., 100.],
+        [-60, -40, 40., 60., 0., 100., "2007-08-01", "2007-09-01"],
+    ]
+    requests["wmo"] = [
+        5900446,
+        5906072,
+        6901929,
+        1900857,
+        3902131,
+        2902696,
+    ]
+
+    def test_chunks_region(self):
+        with argopy.set_options(local_ftp=self.local_ftp):
+            for access_arg in self.requests["region"]:
+                fetcher_args = {"src": self.src, "parallel": True}
+                try:
+                    f = ArgoDataFetcher(**fetcher_args).region(access_arg)
+                    assert is_list_of_strings(f.fetcher.uri)
+                    assert isinstance(f.to_xarray(), xr.Dataset)
+                except Exception:
+                    print("ERROR LOCALFTP request:\n", f.fetcher.uri)
+                    pass
+
+    def test_chunks_wmo(self):
+        with argopy.set_options(local_ftp=self.local_ftp):
+            for access_arg in self.requests["wmo"]:
+                fetcher_args = {"src": self.src, "parallel": True}
+                try:
+                    f = ArgoDataFetcher(**fetcher_args).float(access_arg)
+                    assert is_list_of_strings(f.fetcher.uri)
+                    assert isinstance(f.to_xarray(), xr.Dataset)
+                except Exception:
+                    print("ERROR LOCALFTP request:\n", f.fetcher.uri)
+                    pass
 
 
 if __name__ == '__main__':